use topicform as bouquet edit view
